h06 q01 still not working but DIFFERENTLY

+++ input/prior-values.R
@@ -5,4 +5,4 @@
 values_priors <- tibble(
   model_name = c("prior_1", "prior_0.1", "prior_10", "prior_100"),
   model_name_sym = lapply(model_name, as.symbol),
-  prior_level = as.integer(c(1, 0.1, 10, 100)))
+  prior_level = as.numeric(c(1, 0.1, 10, 100)))
